bugfix RdbBuckets::fastLoad()
If the bigfile was exactly a multple of 4GB it would fail due use of a int32_t instead of int64_t

+++ RdbBuckets.cpp
@@ -2086,7 +2086,7 @@ bool RdbBuckets::fastLoad ( BigFile *f , char* dbname) {
 		return false;
 	}
 
-	int32_t fsize = f->getFileSize();
+	int64_t fsize = f->getFileSize();
 	if ( fsize == 0 ) {
 		return true;
 	}
